study/requirements.txt
raven
pymongo
flask
flask_mail
Flask-BasicAuth
flask_socketio
gevent==1.2.2
kombu
psycopg2-binary
geojson
shapely
requests
werkzeug
geocoder
numpy
sklearn
scipy
nltk
pika
pillow
lxml